seeding in test project seems to work
