+++ packages/server/modules/core/rest/e2e.ts
@@ -1,4 +1,6 @@
+import { seedViewerE2eTestStream } from '@/modules/core/services/e2e'
 import { isDevEnv, isTestEnv } from '@/modules/shared/helpers/envHelper'
+import { ensureError } from '@speckle/shared'
 import * as express from 'express'
 
 module.exports = (app: express.Application) => {
@@ -8,7 +10,14 @@ module.exports = (app: express.Application) => {
    * Set up server seeding endpoints for E2E tests
    */
 
-  app.post('/e2e/seed', async () => {
-    // TODO
+  app.post('/api/e2e/seed', async (_req, res) => {
+    try {
+      const results = await seedViewerE2eTestStream()
+      res.json(results).status(200)
+    } catch (e) {
+      res
+        .json({ error: ensureError(e).message, stack: ensureError(e).stack })
+        .status(500)
+    }
   })
 }
